fix: disable retro tapping for gui mod-taps
Retro tapping ensures that when a mod-tap key is held down and released
without tapping another key, the keycode is sent instead of the
modifier. This is what I want for shift, control and alt, but not for
GUI. By disabling retro tapping for my GUI mod-tap keys I can ensure
that when holding down and releasing any of the GUI keys, the GUI
modifier is sent instead of the keycode.